Bug 1761483 - Fix integer truncation bug in DOM Streams r=smaug
Differential Revision: https://phabricator.services.mozilla.com/D142289

UltraBlame original commit: 03acbef2489bf780ae423f238cc3d6a666195fcf

+++ dom/base/BodyStream.h
@@ -161,8 +161,8 @@ class BodyStream final : public nsIInputStreamCallback,
  private:
   
   void WriteIntoReadRequestBuffer(JSContext* aCx, ReadableStream* aStream,
-                                  JS::Handle<JSObject*> aBuffer, size_t aLength,
-                                  size_t* aByteWritten);
+                                  JS::Handle<JSObject*> aBuffer,
+                                  uint32_t aLength, uint32_t* aByteWritten);
